+++ ext/duckdb/value.c
@@ -16,6 +16,7 @@ static VALUE duckdb_value_s__create_uint32(VALUE klass, VALUE val);
 static VALUE duckdb_value_s__create_uint64(VALUE klass, VALUE val);
 static VALUE duckdb_value_s__create_float(VALUE klass, VALUE val);
 static VALUE duckdb_value_s__create_double(VALUE klass, VALUE val);
+static VALUE duckdb_value_s__create_varchar(VALUE klass, VALUE str);
 static VALUE duckdb_value_s_create_null(VALUE klass);
 
 static const rb_data_type_t value_data_type = {
@@ -92,6 +93,13 @@ static VALUE duckdb_value_s__create_float(VALUE klass, VALUE val) {
 
 static VALUE duckdb_value_s__create_double(VALUE klass, VALUE val) {
     duckdb_value value = duckdb_create_double(NUM2DBL(val));
+    return rbduckdb_value_new(value);
+}
+
+static VALUE duckdb_value_s__create_varchar(VALUE klass, VALUE str) {
+    const char *str_ptr = StringValuePtr(str);
+    idx_t str_len = RSTRING_LEN(str);
+    duckdb_value value = duckdb_create_varchar_length(str_ptr, str_len);
     return rbduckdb_value_new(value);
 }
 
@@ -236,6 +244,7 @@ void rbduckdb_init_duckdb_value(void) {
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_uint64", duckdb_value_s__create_uint64, 1);
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_float", duckdb_value_s__create_float, 1);
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_double", duckdb_value_s__create_double, 1);
+    rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_varchar", duckdb_value_s__create_varchar, 1);
     rb_define_singleton_method(cDuckDBValue, "create_null", duckdb_value_s_create_null, 0);
 }