+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - add `DuckDB::Value.create_null`.
 - add `DuckDB::Value.create_float`.
 - add `DuckDB::Value.create_double`.
+- add `DuckDB::Value.create_varchar`.
 
 ## Breaking Changes
 - rename `DuckDB::ValueImpl` to `DuckDB::Value`